Beginnings of rndc support; switch to Debian
Switched from Ubuntu to Debian for the base image (~100MB image size
reduction).

The rndc key is now copied to /rndc on startup. This may be mounted as a
volume to allow access to it from outside the container. Port 953 is now
listed as an EXPOSE, too, but named.conf doesn't allow external control
(yet).

Zones are now read from a /zones directory. The contents of this repo's
zones directory are COPYed to it on build, and the named.conf.local file
is generated on startup, listing each zone file in /zones in a zone {};
statement. The filename must be a legal BIND zone (e.g.,
168.192.in-addr.arpa. or mydomain.com.). If you bind-mount /zones (it's
listed as VOLUME), you need to make sure your zones are in the directory
you're mounting. This, coupled with the rndc work above, will allow
modification of zones in the running container.

Dynamic DNS is not yet supported.